tidying up the interface

+++ src/api/crudServices.ts
@@ -8,28 +8,14 @@ import {
 } from 'firebase/firestore';
 import db from './firebaseServices';
 
-interface Data {
+interface TaskData {
   taskName: string;
   taskType: string;
   taskDate: string;
   taskTime: number;
-}
-
-interface DataDelete {
-  taskName: string;
-  taskType: string;
-  taskTime: number;
-  id: string;
-}
-
-interface DataUpdate {
-  taskName: string;
-  taskType: string;
-  taskTime: number;
-  startTime: string;
-  endTime: string;
-  taskDate: string;
-  id: string;
+  id?: string;
+  startTime?: string;
+  endTime?: string;
 }
 
 export async function getData() {
@@ -37,12 +23,12 @@ export async function getData() {
   const myDocuments = await getDocs(myCollection);
 
   const result = myDocuments.docs.map((doc) => {
-    return { id: doc.id, ...doc.data() };
+    return { id: doc.id, ...doc.data() } as TaskData;
   });
   return result;
 }
 
-export async function addData(data: Data) {
+export async function addData(data: TaskData) {
   try {
     const myCollection = collection(db, 'tasks');
     await addDoc(myCollection, data);
@@ -51,37 +37,43 @@ export async function addData(data: Data) {
   }
 }
 
-export async function deleteRecord(record: DataDelete) {
-  const myCollection = collection(db, 'tasks');
-  const docRef = await doc(myCollection, record.id);
-  deleteDoc(docRef);
-  console.log('Dokument deleted with ID:', docRef.id);
+export async function deleteRecord(record: TaskData) {
+  if (record.id) {
+    const myCollection = collection(db, 'tasks');
+    const docRef = doc(myCollection, record.id);
+    await deleteDoc(docRef);
+    console.log('Dokument deleted with ID:', record.id);
+  }
 }
 
 export const fetchExistingTaskTypes = async () => {
   try {
-    const myCollection = collection(db, "tasks");
+    const myCollection = collection(db, 'tasks');
     const myDocuments = await getDocs(myCollection);
 
     const taskTypes = myDocuments.docs.map((doc) => doc.data().taskType);
     return taskTypes;
   } catch (error) {
-    console.error("Wystąpił błąd podczas pobierania istniejących rodzajów zadań:", error);
+    console.error('Wystąpił błąd podczas pobierania istniejących rodzajów zadań:', error);
     return [];
   }
 };
-export async function updateRecord(record: DataUpdate) {
-  const myCollection = collection(db, 'tasks');
-  const docRef = await doc(myCollection, record.id);
 
-  updateDoc(docRef, {
-    taskName: record.taskName,
-    taskType: record.taskType,
-    taskTime: record.taskTime,
-    startTime: record.startTime,
-    endTime: record.endTime,
-    taskDate: record.taskDate,
-  });
+export async function updateRecord(record: TaskData) {
+  if (record.id) {
+    const myCollection = collection(db, 'tasks');
+    const docRef = doc(myCollection, record.id);
 
-  console.log("Dokument updated with ID:", docRef.id);
+    const updatedData: Partial<TaskData> = {
+      taskName: record.taskName,
+      taskType: record.taskType,
+      taskTime: record.taskTime,
+      taskDate: record.taskDate,
+      startTime: record.startTime,
+      endTime: record.endTime,
+    };
+
+    await updateDoc(docRef, updatedData);
+    console.log('Dokument updated with ID:', record.id);
+  }
 }
